--- src/main/java/com/kdt/wolf/domain/user/service/AlertService.java
@@ -1,11 +1,0 @@
-package com.kdt.wolf.domain.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AlertService {
-
-
-}
