+++ src/main/java/com/kdt/wolf/domain/user/service/UserService.java
@@ -1,8 +1,6 @@
 package com.kdt.wolf.domain.user.service;
 
-import com.kdt.wolf.domain.user.dao.AlertDao;
 import com.kdt.wolf.domain.user.dao.UserDao;
-import com.kdt.wolf.domain.user.dto.FcmDto.AlertDto;
 import com.kdt.wolf.domain.user.dto.SignUpDto.SignUpRequest;
 import com.kdt.wolf.domain.user.dto.UserDto.UserProfileDetailResponse;
 import com.kdt.wolf.domain.user.dto.UserDto.UserProfileResponse;
@@ -11,7 +9,6 @@ import com.kdt.wolf.domain.user.entity.UserEntity;
 import com.kdt.wolf.global.exception.BusinessException;
 import com.kdt.wolf.global.exception.code.ExceptionCode;
 import jakarta.transaction.Transactional;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +16,6 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class UserService {
     private final UserDao userDao;
-    private final AlertDao alertDao;
 
     public UserProfileDetailResponse getUserProfileDetail(Long userId) {
         UserEntity user = userDao.findById(userId);
@@ -43,13 +39,5 @@ public class UserService {
         UserEntity user = userDao.findById(userId);
         userDao.updateUser(user.updateProfile(request));
         return user.toUserProfileDetailResponse();
-    }
-
-    public List<AlertDto> getAlarms(Long userId) {
-        return alertDao.getAlarms(userId);
-    }
-
-    public List<AlertDto> getAlarmsPreview(Long userId) {
-        return alertDao.getAlarmsPreview(userId);
     }
 }
