+++ src/main/java/com/kdt/wolf/global/fcm/service/FcmService.java
@@ -1,8 +1,9 @@
 package com.kdt.wolf.global.fcm.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class FcmService {
-
 }
